# improve the performance of gc_content(::Kmer)