Before (master):
```
julia> @benchmark gc_content($(DNAKmer("ACGT"^8)))
BenchmarkTools.Trial:
  samples:          10000
  evals/sample:     992
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     37.00 ns (0.00% GC)
  median time:      47.00 ns (0.00% GC)
  mean time:        47.93 ns (0.00% GC)
  maximum time:     1.94 μs (0.00% GC)

```

After:
```
julia> @benchmark gc_content($(DNAKmer("ACGT"^8)))
BenchmarkTools.Trial:
  samples:          10000
  evals/sample:     1000
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  0.00 bytes
  allocs estimate:  0
  minimum time:     6.00 ns (0.00% GC)
  median time:      6.00 ns (0.00% GC)
  mean time:        6.02 ns (0.00% GC)
  maximum time:     60.00 ns (0.00% GC)

```